Fix a comparison between a size_t and -1 by casting the int to a size_t. This comparison was causing a warning with the MTA compiler.
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12308 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/src/chplcopygc.c
@@ -41,7 +41,7 @@ void _chpl_gc_copy_collect(void) {
   scanptr = _to_space->head;
   while (scanptr != _to_space->current) {
     size_t *offsets = cid2offsets(*(_class_id*)scanptr);
-    if (offsets[1] == -1) {
+    if (offsets[1] == (size_t)-1) {
       // data class (array)
       int i;
       // Bad, could be uint(64) index size
